requirements.txt
uvicorn==0.26.0
marvin==2.1.3
openai==1.8.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
fastapi==0.109.0
python-dotenv==1.0.1
numpy 
pandas
pytest